update 23 checks existing files and updates every 10 sec

+++ 23 - Image Board Media Downloader/main.py
@@ -1,25 +1,31 @@
-import requests, sys, urllib.request, os
+import requests, sys, urllib.request, os, time
 from bs4 import BeautifulSoup
 
 url = str(sys.argv[1]).replace('http://', '').replace('https://', '')
 
-data = requests.get(f'https://{url}')
+start_time = time.time()
+interval = 10
+for x in range(20):
+    time.sleep(start_time + x * interval - time.time())
 
-soup = BeautifulSoup(data.text, 'html.parser')
+    data = requests.get(f'https://{url}')
 
-imgs = []
-for i in soup.find_all('a', {'class': 'fileThumb'}, href=True):
-    imgs.append(i['href'].replace('//', ''))
+    soup = BeautifulSoup(data.text, 'html.parser')
 
-names = []
-for i in soup.find_all('div', {'class': 'fileText'}):
-    for x in i.find_all('a'):
-        names.append(x.text)
+    imgs = []
+    for i in soup.find_all('a', {'class': 'fileThumb'}, href=True):
+        imgs.append(i['href'].replace('//', ''))
 
-if not os.path.exists(f'downloads/{url.split("/")[1]} - {url.split("/")[3]}'):
-    os.makedirs(f'downloads/{url.split("/")[1]} - {url.split("/")[3]}')
+    names = []
+    for i in soup.find_all('div', {'class': 'fileText'}):
+        for x in i.find_all('a'):
+            names.append(x.text)
 
-for i in range(0, len(imgs)):
-    print(names[i], end=' - ')
-    urllib.request.urlretrieve(f'https://{imgs[i]}', f'downloads/{url.split("/")[1]} - {url.split("/")[3]}/{names[i]}')
-    print("Done")
+    if not os.path.exists(f'downloads/{url.split("/")[1]} - {url.split("/")[3]}'):
+        os.makedirs(f'downloads/{url.split("/")[1]} - {url.split("/")[3]}')
+
+    for i in range(0, len(imgs)):
+        if(os.path.isfile(f'downloads/{url.split("/")[1]} - {url.split("/")[3]}/{names[i]}') == False):
+            print(names[i], end=' - ')
+            urllib.request.urlretrieve(f'https://{imgs[i]}', f'downloads/{url.split("/")[1]} - {url.split("/")[3]}/{names[i]}')
+            print("Done")
